feat: migrate link component to tsx
